pyproject.toml
[tool.poetry]
name = "realmate-challenge"
version = "0.1.0"
description = ""
authors = ["Jean Landim"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
djangorestframework = "^3.16.0"
celery = "^5.5.3"
redis = "^6.2.0"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
python-dotenv = "^1.1.0"
ruff = "^0.12.0"
pytest = "^8.4.0"
pytest-cov = "^6.2.1"
pytest-django = "^4.11.1"
pyyaml = "5.3"
pytest-docker = "^3.2.2"
loguru = "^0.7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py310"
include = ["**/*.py"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false
